interpolation works

+++ webapp/src/test/test/org/nees/mustsim/replay/test/utils/mgen/GenerateActualEmpty.java
@@ -12,8 +12,9 @@ public class GenerateActualEmpty extends GenerateActualColumn {
 		}
 	}
 
-	public GenerateActualEmpty(int rowSize, int increment) {
-		super(rowSize, increment,0);
+	public GenerateActualEmpty(int rowSize, double increment) {
+		super(rowSize, increment, 0);
 	}
+
 
 }
